Potential fix for code scanning alert no. 3: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger integer type to avoid overflow. This can be achieved by casting one of the operands to the larger type before performing the multiplication. Specifically, we should cast dims[i].is to INT before multiplying it by is.

+++ libfftw-3.3.8/api/mktensor-iodims.h
@@ -28,8 +28,8 @@ tensor *MKTENSOR_IODIMS(int rank, const IODIM *dims, int is, int os)
      if (FINITE_RNK(rank)) {
           for (i = 0; i < rank; ++i) {
                x->dims[i].n = dims[i].n;
-               x->dims[i].is = dims[i].is * is;
-               x->dims[i].os = dims[i].os * os;
+               x->dims[i].is = (INT)dims[i].is * is;
+               x->dims[i].os = (INT)dims[i].os * os;
           }
      }
      return x;
